requirements.txt
Django==1.11.18
google_api_python_client==1.7.8
djangorestframework==3.9.4
